initial work on support showing a split view of the lot info and the lots
